fix: remove .int() chain incompatible with Zod v4 in reminders-section

+++ components/parametres/reminders-section.tsx
@@ -10,8 +10,8 @@ import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 
 const schema = z.object({
-  reminder_delay_j7: z.coerce.number().int().min(1).max(30),
-  reminder_delay_j14: z.coerce.number().int().min(1).max(60),
+  reminder_delay_j7: z.coerce.number().min(1).max(30),
+  reminder_delay_j14: z.coerce.number().min(1).max(60),
 })
 
 type FormValues = z.infer<typeof schema>
